Removed unused imports and moved encryption provider to util package
