+++ lib/widget/stateless/main.dart
@@ -7,7 +7,6 @@ import 'user_home.dart';
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
